requirements.txt
yandex-music>=2.1.1
asyncio>=3.4.3
aiogram>=3.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
